requirements.txt
psycopg2-binary==2.9.9
faker==24.9.0
pandas==2.2.2
streamlit==1.35.0
plotly==5.22.0
dbt-postgres==1.8.2
apache-airflow==2.9.1
